+++ packages/vm/src/imports.rs
@@ -74,20 +74,20 @@ pub fn native_db_remove<A: BackendApi, S: Storage, Q: Querier>(
     do_remove(env, key_ptr)
 }
 
-pub fn native_canonicalize_address<A: BackendApi, S: Storage, Q: Querier>(
+pub fn native_addr_canonicalize<A: BackendApi, S: Storage, Q: Querier>(
     env: &Environment<A, S, Q>,
     source_ptr: u32,
     destination_ptr: u32,
 ) -> VmResult<u32> {
-    do_canonicalize_address(&env, source_ptr, destination_ptr)
+    do_addr_canonicalize(&env, source_ptr, destination_ptr)
 }
 
-pub fn native_humanize_address<A: BackendApi, S: Storage, Q: Querier>(
+pub fn native_addr_humanize<A: BackendApi, S: Storage, Q: Querier>(
     env: &Environment<A, S, Q>,
     source_ptr: u32,
     destination_ptr: u32,
 ) -> VmResult<u32> {
-    do_humanize_address(&env, source_ptr, destination_ptr)
+    do_addr_humanize(&env, source_ptr, destination_ptr)
 }
 
 pub fn native_secp256k1_verify<A: BackendApi, S: Storage, Q: Querier>(
@@ -226,7 +226,7 @@ fn do_remove<A: BackendApi, S: Storage, Q: Querier>(
     Ok(())
 }
 
-fn do_canonicalize_address<A: BackendApi, S: Storage, Q: Querier>(
+fn do_addr_canonicalize<A: BackendApi, S: Storage, Q: Querier>(
     env: &Environment<A, S, Q>,
     source_ptr: u32,
     destination_ptr: u32,
@@ -261,7 +261,7 @@ fn do_canonicalize_address<A: BackendApi, S: Storage, Q: Querier>(
     }
 }
 
-fn do_humanize_address<A: BackendApi, S: Storage, Q: Querier>(
+fn do_addr_humanize<A: BackendApi, S: Storage, Q: Querier>(
     env: &Environment<A, S, Q>,
     source_ptr: u32,
     destination_ptr: u32,
@@ -569,8 +569,8 @@ mod tests {
                 "db_scan" => Function::new_native(&store, |_a: u32, _b: u32, _c: i32| -> u32 { 0 }),
                 "db_next" => Function::new_native(&store, |_a: u32| -> u32 { 0 }),
                 "query_chain" => Function::new_native(&store, |_a: u32| -> u32 { 0 }),
-                "canonicalize_address" => Function::new_native(&store, |_a: u32, _b: u32| -> u32 { 0 }),
-                "humanize_address" => Function::new_native(&store, |_a: u32, _b: u32| -> u32 { 0 }),
+                "addr_canonicalize" => Function::new_native(&store, |_a: u32, _b: u32| -> u32 { 0 }),
+                "addr_humanize" => Function::new_native(&store, |_a: u32, _b: u32| -> u32 { 0 }),
                 "secp256k1_verify" => Function::new_native(&store, |_a: u32, _b: u32, _c: u32| -> u32 { 0 }),
                 "secp256k1_recover_pubkey" => Function::new_native(&store, |_a: u32, _b: u32, _c: u32| -> u64 { 0 }),
                 "ed25519_verify" => Function::new_native(&store, |_a: u32, _b: u32, _c: u32| -> u32 { 0 }),
@@ -897,7 +897,7 @@ mod tests {
     }
 
     #[test]
-    fn do_canonicalize_address_works() {
+    fn do_addr_canonicalize_works() {
         let api = MockApi::default();
         let (env, mut instance) = make_instance(api);
         let api = MockApi::default();
@@ -908,13 +908,13 @@ mod tests {
         leave_default_data(&env);
 
         let api = MockApi::default();
-        do_canonicalize_address::<MA, MS, MQ>(&env, source_ptr, dest_ptr).unwrap();
+        do_addr_canonicalize::<MA, MS, MQ>(&env, source_ptr, dest_ptr).unwrap();
         let data = force_read(&env, dest_ptr);
         assert_eq!(data.len(), api.canonical_length);
     }
 
     #[test]
-    fn do_canonicalize_address_reports_invalid_input_back_to_contract() {
+    fn do_addr_canonicalize_reports_invalid_input_back_to_contract() {
         let api = MockApi::default();
         let (env, mut instance) = make_instance(api);
 
@@ -925,24 +925,24 @@ mod tests {
 
         leave_default_data(&env);
 
-        let res = do_canonicalize_address::<MA, MS, MQ>(&env, source_ptr1, dest_ptr).unwrap();
+        let res = do_addr_canonicalize::<MA, MS, MQ>(&env, source_ptr1, dest_ptr).unwrap();
         assert_ne!(res, 0);
         let err = String::from_utf8(force_read(&env, res)).unwrap();
         assert_eq!(err, "Input is not valid UTF-8");
 
-        let res = do_canonicalize_address::<MA, MS, MQ>(&env, source_ptr2, dest_ptr).unwrap();
+        let res = do_addr_canonicalize::<MA, MS, MQ>(&env, source_ptr2, dest_ptr).unwrap();
         assert_ne!(res, 0);
         let err = String::from_utf8(force_read(&env, res)).unwrap();
         assert_eq!(err, "Input is empty");
 
-        let res = do_canonicalize_address::<MA, MS, MQ>(&env, source_ptr3, dest_ptr).unwrap();
+        let res = do_addr_canonicalize::<MA, MS, MQ>(&env, source_ptr3, dest_ptr).unwrap();
         assert_ne!(res, 0);
         let err = String::from_utf8(force_read(&env, res)).unwrap();
         assert_eq!(err, "Invalid input: human address too long");
     }
 
     #[test]
-    fn do_canonicalize_address_fails_for_broken_backend() {
+    fn do_addr_canonicalize_fails_for_broken_backend() {
         let api = MockApi::new_failing("Temporarily unavailable");
         let (env, mut instance) = make_instance(api);
 
@@ -951,7 +951,7 @@ mod tests {
 
         leave_default_data(&env);
 
-        let result = do_canonicalize_address::<MA, MS, MQ>(&env, source_ptr, dest_ptr);
+        let result = do_addr_canonicalize::<MA, MS, MQ>(&env, source_ptr, dest_ptr);
         match result.unwrap_err() {
             VmError::BackendErr {
                 source: BackendError::Unknown { msg, .. },
@@ -964,7 +964,7 @@ mod tests {
     }
 
     #[test]
-    fn do_canonicalize_address_fails_for_large_inputs() {
+    fn do_addr_canonicalize_fails_for_large_inputs() {
         let api = MockApi::default();
         let (env, mut instance) = make_instance(api);
 
@@ -973,7 +973,7 @@ mod tests {
 
         leave_default_data(&env);
 
-        let result = do_canonicalize_address::<MA, MS, MQ>(&env, source_ptr, dest_ptr);
+        let result = do_addr_canonicalize::<MA, MS, MQ>(&env, source_ptr, dest_ptr);
         match result.unwrap_err() {
             VmError::CommunicationErr {
                 source:
@@ -990,7 +990,7 @@ mod tests {
     }
 
     #[test]
-    fn do_canonicalize_address_fails_for_small_destination_region() {
+    fn do_addr_canonicalize_fails_for_small_destination_region() {
         let api = MockApi::default();
         let (env, mut instance) = make_instance(api);
 
@@ -999,7 +999,7 @@ mod tests {
 
         leave_default_data(&env);
 
-        let result = do_canonicalize_address::<MA, MS, MQ>(&env, source_ptr, dest_ptr);
+        let result = do_addr_canonicalize::<MA, MS, MQ>(&env, source_ptr, dest_ptr);
         match result.unwrap_err() {
             VmError::CommunicationErr {
                 source: CommunicationError::RegionTooSmall { size, required, .. },
@@ -1013,7 +1013,7 @@ mod tests {
     }
 
     #[test]
-    fn do_humanize_address_works() {
+    fn do_addr_humanize_works() {
         let api = MockApi::default();
         let (env, mut instance) = make_instance(api);
         let api = MockApi::default();
@@ -1024,13 +1024,13 @@ mod tests {
 
         leave_default_data(&env);
 
-        let error_ptr = do_humanize_address::<MA, MS, MQ>(&env, source_ptr, dest_ptr).unwrap();
+        let error_ptr = do_addr_humanize::<MA, MS, MQ>(&env, source_ptr, dest_ptr).unwrap();
         assert_eq!(error_ptr, 0);
         assert_eq!(force_read(&env, dest_ptr), source_data);
     }
 
     #[test]
-    fn do_humanize_address_reports_invalid_input_back_to_contract() {
+    fn do_addr_humanize_reports_invalid_input_back_to_contract() {
         let api = MockApi::default();
         let (env, mut instance) = make_instance(api);
 
@@ -1039,14 +1039,14 @@ mod tests {
 
         leave_default_data(&env);
 
-        let res = do_humanize_address::<MA, MS, MQ>(&env, source_ptr, dest_ptr).unwrap();
+        let res = do_addr_humanize::<MA, MS, MQ>(&env, source_ptr, dest_ptr).unwrap();
         assert_ne!(res, 0);
         let err = String::from_utf8(force_read(&env, res)).unwrap();
         assert_eq!(err, "Invalid input: canonical address length not correct");
     }
 
     #[test]
-    fn do_humanize_address_fails_for_broken_backend() {
+    fn do_addr_humanize_fails_for_broken_backend() {
         let api = MockApi::new_failing("Temporarily unavailable");
         let (env, mut instance) = make_instance(api);
 
@@ -1055,7 +1055,7 @@ mod tests {
 
         leave_default_data(&env);
 
-        let result = do_humanize_address::<MA, MS, MQ>(&env, source_ptr, dest_ptr);
+        let result = do_addr_humanize::<MA, MS, MQ>(&env, source_ptr, dest_ptr);
         match result.unwrap_err() {
             VmError::BackendErr {
                 source: BackendError::Unknown { msg, .. },
@@ -1066,7 +1066,7 @@ mod tests {
     }
 
     #[test]
-    fn do_humanize_address_fails_for_input_too_long() {
+    fn do_addr_humanize_fails_for_input_too_long() {
         let api = MockApi::default();
         let (env, mut instance) = make_instance(api);
 
@@ -1075,7 +1075,7 @@ mod tests {
 
         leave_default_data(&env);
 
-        let result = do_humanize_address::<MA, MS, MQ>(&env, source_ptr, dest_ptr);
+        let result = do_addr_humanize::<MA, MS, MQ>(&env, source_ptr, dest_ptr);
         match result.unwrap_err() {
             VmError::CommunicationErr {
                 source:
@@ -1092,7 +1092,7 @@ mod tests {
     }
 
     #[test]
-    fn do_humanize_address_fails_for_destination_region_too_small() {
+    fn do_addr_humanize_fails_for_destination_region_too_small() {
         let api = MockApi::default();
         let (env, mut instance) = make_instance(api);
         let api = MockApi::default();
@@ -1103,7 +1103,7 @@ mod tests {
 
         leave_default_data(&env);
 
-        let result = do_humanize_address::<MA, MS, MQ>(&env, source_ptr, dest_ptr);
+        let result = do_addr_humanize::<MA, MS, MQ>(&env, source_ptr, dest_ptr);
         match result.unwrap_err() {
             VmError::CommunicationErr {
                 source: CommunicationError::RegionTooSmall { size, required, .. },